+++ includes/class-magick-mixture.php
@@ -129,6 +129,8 @@ class Magick_Mixture
      */
     public function run()
     {
+        add_action('init', array('MaBox_Block_Patterns', 'register'));
+
         //对js文件进行module接入
         add_filter('script_loader_tag', array(__CLASS__, 'refund_type_script'), 10, 2);
     }